python-voice-agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-groq>=0.6.0
livekit-plugins-elevenlabs>=0.6.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0
